requirements.txt
configargparse==1.7.1
einops==0.8.1
fpsample==0.3.3
matplotlib==3.10.6
numpy==2.1.2
pillow==11.0.0
plyfile==1.1.2
pycolmap==3.12.0
pyyaml==6.0.2
scipy==1.16.2
torch==2.8.0+cu129
tensorboard==2.20.0
tqdm==4.67.1
warp-lang==1.9.0
open3d==0.19.0